remove debugStack stuff in populate since these aren't used
